pyproject.toml
[project]
name = "wolfbot3"
version = "0.1.0"
description = "Bot for helping moderate social deduction games on Discord"
authors = [
    {name = "Tyler Jackson", email = "tylerj714@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.14"
dependencies = [
    "discord.py (>=2.5.0)",
    "python-dotenv (>=1.1.1)",
    "pydantic (>=2.11.7)"
]

[tool.poetry]
packages = [{include = "bot"}]

[build-system]
requires = ["poetry-core >= 2.0.0, <3.0.0"]
build-backend = "poetry.core.masonry.api"